address @ashkitten's comments
